agent-runner/requirements.txt
# HTTP client — used by AgentClient and Slack delivery
httpx>=0.27,<1.0

# Timezone-aware datetime arithmetic
pytz>=2024.1

# Postgres client — used for job-run state and audit log
# psycopg2-binary ships the shared library; use psycopg2 for compile-from-source envs.
psycopg2-binary>=2.9,<3.0

# Environment variable loading from .env files (development convenience)
python-dotenv>=1.0,<2.0
